middleware/requirements.txt
duckdb-engine
fastapi
starlette-graphene3
graphene_sqlalchemy
python-dotenv